reafactor(sidebar): divide bloat sidebar into SidebarHeader, SidebarBody, SidebarFooter and SidebarMobile components
