+++ components/sidebar/mobile-sidebar.tsx
@@ -1,0 +1,68 @@
+import { Menu } from "lucide-react"
+
+import { Button } from "@/components/ui/button"
+import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet"
+import {
+  BarChart3,
+  FileText,
+  LayoutDashboard,
+  Settings,
+  Users,
+} from "lucide-react"
+import { usePathname } from "next/navigation"
+import Link from "next/link"
+
+const menuItems = [
+  { href: "/dashboard", label: "Dashboard", icon: LayoutDashboard },
+  { href: "/dashboard/analytics", label: "Analytics", icon: BarChart3 },
+  { href: "/dashboard/users", label: "Users", icon: Users },
+  { href: "/dashboard/reports", label: "Reports", icon: FileText },
+  { href: "/dashboard/settings", label: "Settings", icon: Settings },
+]
+
+const MobileSidebar = () => {
+  const pathname = usePathname()
+
+  return (
+    <Sheet>
+      <SheetTrigger asChild>
+        <Button
+          variant="outline"
+          size="icon"
+          className="fixed left-4 top-4 z-50 md:hidden"
+        >
+          <Menu className="h-5 w-5" />
+        </Button>
+      </SheetTrigger>
+      <SheetContent side="left" className="w-64 p-0">
+        <div className="flex h-full flex-col">
+          <div className="flex h-16 items-center border-b px-6">
+            <h2 className="text-lg font-semibold">Acme Dashboard</h2>
+          </div>
+          <nav className="flex-1 space-y-1 p-4">
+            {menuItems.map((item) => {
+              const Icon = item.icon
+              const isActive = pathname === item.href
+              return (
+                <Link
+                  key={item.href}
+                  href={item.href}
+                  className={`flex items-center gap-3 rounded-lg px-3 py-2 text-sm font-medium transition-colors ${
+                    isActive
+                      ? "bg-primary text-primary-foreground"
+                      : "hover:bg-muted"
+                  }`}
+                >
+                  <Icon className="h-5 w-5" />
+                  <span>{item.label}</span>
+                </Link>
+              )
+            })}
+          </nav>
+        </div>
+      </SheetContent>
+    </Sheet>
+  )
+}
+
+export default MobileSidebar
